+++ src/ch5/2d-incremental-collision-checking/file_loader.py
