+++ src/ch5/2d-incremental-collision-checking/for-later/polygon-driver.py
