+++ src/ch5/2d-incremental-collision-checking/json-polygons/displaced_rectangle.json
