+++ src/ch5/2d-incremental-collision-checking/json-polygons/octagon.json
