--- src/ch5/2d-incremental-collision-checking/points.txt
@@ -1,3 +1,0 @@
-0,0
-0,1
-0,2
