+++ src/ch5/2d-incremental-collision-checking/pygame_rendering/pygame_loop_support.py
