+++ src/ch5/2d-incremental-collision-checking/pygame_rendering/render_support.py
@@ -2,6 +2,7 @@
 
 import pygame
 import time
+
 colors = {
   "black" : (0,0,0),
   "indigo" : (48,79,254),
