+++ src/ch5/2d-incremental-collision-checking/support/doubly_connected_edge_list.py
